requirements.txt
streamlit==1.22.0
azure-identity==1.13.0
azure-mgmt-resource==23.0.1
azure-mgmt-cosmosdb==9.2.0
azure-cosmos==4.3.1